+++ src/engine/index.ts
@@ -22,3 +22,4 @@ export * from './hotspots';
 export * from './appExport';
 export * from './projectFile';
 export * from './svgExport';
+export * from './dataPlot';